fix: Fix more "Returning Any from functions" errors in analysis and DSP (173 → 158 errors)
- feature_extractor.py: Wrap float returns with float() (3 fixes)
- simd_accelerator.py: Wrap array returns with np.asarray() (3 fixes)
- memory_pool.py: Wrap buffer returns with np.asarray() (3 fixes)
- parameter_generator.py: Wrap np.clip() returns with float() (5 fixes)
- adaptive.py: Wrap np.clip() returns with float() (2 fixes)
- hybrid_mode.py: Cast adaptive_processor return to np.ndarray (1 fix)

Total this batch: 15 errors fixed (173 → 158). Cumulative: 28 errors fixed from original 186 (84.8% improvement).

+++ auralis/analysis/ml/feature_extractor.py
@@ -130,7 +130,7 @@ class FeatureExtractor:
             fallback=0.0
         ))
 
-        return bandwidth
+        return float(bandwidth)
 
     def _spectral_contrast(self, audio: np.ndarray, n_bands: int = 6) -> List[float]:
         """Calculate spectral contrast across frequency bands"""
@@ -162,7 +162,7 @@ class FeatureExtractor:
         geometric_mean = np.exp(np.mean(np.log(magnitude)))
         arithmetic_mean = np.mean(magnitude)
 
-        return geometric_mean / arithmetic_mean
+        return float(geometric_mean / arithmetic_mean)
 
     def _tempo_stability(self, audio: np.ndarray) -> float:
         """Estimate tempo stability"""
@@ -212,7 +212,7 @@ class FeatureExtractor:
         peak_energy = np.sum(autocorr[peaks])
         total_energy = np.sum(autocorr)
 
-        return peak_energy / (total_energy + 1e-10)
+        return float(peak_energy / (total_energy + 1e-10))
 
     def _fundamental_frequency(self, audio: np.ndarray) -> float:
         """Estimate fundamental frequency"""
